Linux compile problem elimination
git-svn-id: file:///local/rexx/oorexx/oorexx.svn@1466 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ sandbox/rick/nolong/kernel/parser/Scanner.cpp
@@ -755,7 +755,7 @@ RexxToken *RexxSource::sourceNextToken(
         value = this->commonString(value);
                                        /* record current position in clause */
         this->clause->setEnd(this->line_number, this->line_offset);
-        if (length > MAX_SYMBOL_LENGTH)/* result too long?                  */
+        if (length > (size_t)MAX_SYMBOL_LENGTH)/* result too long?                  */
                                        /* report the error                  */
           report_error1(Error_Name_too_long_name, value);
         inch = this->current[start];   /* get the first character           */
